feat: impl vault sync connector, define common abstractions.
