--- async/vault-async/src/main/java/co/com/bancolombia/secretsmanager/config/HttpProperties.java
@@ -1,16 +1,0 @@
-package co.com.bancolombia.secretsmanager.config;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class HttpProperties {
-
-    @Builder.Default
-    private int connectionTimeout = 5; //in seconds
-}
