--- async/vault-async/src/main/java/co/com/bancolombia/secretsmanager/connector/secret/SecretPayload.java
@@ -1,15 +1,0 @@
-package co.com.bancolombia.secretsmanager.connector.secret;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class SecretPayload {
-    private Map<String, Object> data;
-    private Map<String, Object> metadata;
-}
