+++ async/vault-async/vault-async.gradle
@@ -1,15 +1,15 @@
 dependencies {
     api project(":secrets-manager-api")
+    api project(":vault-commons")
     implementation "io.projectreactor.addons:reactor-extra:${reactorExtraVersion}"
     implementation "io.projectreactor:reactor-core:${reactorCoreVersion}"
     implementation "com.google.code.gson:gson:${gsonVersion}"
     implementation "com.github.ben-manes.caffeine:caffeine:${cafeineVersion}"
     testImplementation "io.projectreactor:reactor-test:${reactorCoreVersion}"
     testImplementation("com.squareup.okhttp3:mockwebserver:4.9.3")
-
 }
 
 ext {
     artifactId = 'vault-async'
-    artifactDescription = 'Secrets Manager'
+    artifactDescription = 'Secrets Manager async connector for Vault'
 }
