+++ gradle.properties
@@ -1,4 +1,4 @@
-version=4.3.1
+version=4.4.0
 springBootVersion=3.2.1
 reactorCoreVersion=3.6.1
 reactorExtraVersion=3.5.1
