+++ sync/vault-commons/src/main/java/co/com/bancolombia/secretsmanager/vault/auth/K8sAuth.java
@@ -1,0 +1,15 @@
+package co.com.bancolombia.secretsmanager.vault.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+public class K8sAuth {
+    private String jwt;
+    private String role;
+}
