+++ sync/vault-commons/src/main/java/co/com/bancolombia/secretsmanager/vault/secret/SecretPayload.java
@@ -1,0 +1,15 @@
+package co.com.bancolombia.secretsmanager.vault.secret;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SecretPayload {
+    private Map<String, Object> data;
+    private Map<String, Object> metadata;
+}
