+++ sync/vault-commons/src/main/java/co/com/bancolombia/secretsmanager/vault/secret/SecretResponse.java
@@ -1,0 +1,23 @@
+package co.com.bancolombia.secretsmanager.vault.secret;
+
+import com.google.gson.annotations.SerializedName;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SecretResponse {
+
+    @SerializedName("request_id")
+    private String requestId;
+
+    @SerializedName("lease_id")
+    private String leaseId;
+    private boolean renewable;
+
+    @SerializedName("lease_duration")
+    private int leaseDuration;
+    private SecretPayload data;
+}
