+++ sync/vault-commons/src/test/resources/certificate.arm
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDSDCCAjACAQEwDQYJKoZIhvcNAQELBQAwajELMAkGA1UEBhMCQ08xDDAKBgNV
+BAgMA0FOVDERMA8GA1UEBwwITUVERUxMSU4xEjAQBgNVBAoMCWxvY2FsaG9zdDES
+MBAGA1UECwwJbG9jYWxob3N0MRIwEAYDVQQDDAlsb2NhbGhvc3QwHhcNMjMxMjIy
+MDI1NzUxWhcNMjQwMzIxMDI1NzUxWjBqMQswCQYDVQQGEwJDTzEMMAoGA1UECAwD
+QU5UMREwDwYDVQQHDAhNRURFTExJTjESMBAGA1UECgwJbG9jYWxob3N0MRIwEAYD
+VQQLDAlsb2NhbGhvc3QxEjAQBgNVBAMMCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcN
+AQEBBQADggEPADCCAQoCggEBALlAo5u+D/fD+v5A4oXcF/14m15mYX+cx0svOKBD
+dNigECPu4UghvmCia3pGFyRP72P1Cx3jye6ImLRCNL7JJFV7BuOiMk1OP+BkqAG0
+RNNt2okMsngkmHTyIHcBdu7QtNWewaZN7WRLa0cOgYMeKc1orftBNNfe7LpLtjuy
+wqzrqi3qI2Y2xXdjTWUVySXzDfVSCnLS+mYNHF5XEkqZXwA4j+rJN9ipeWXxm78G
+sft603a+u+f7PwvPHzMltHI7OxuaJpPC9KzUzgeJ9vzCMz/9Klmp/VrDjigqYMnW
+j3EB5eh7iXamJslexDotHZ2q6pQOq2OAPOVwsExPWc/ILkMCAwEAATANBgkqhkiG
+9w0BAQsFAAOCAQEAplB1L6fxpoQN0ynZdfRaXblvwLjzjVZe0j+p120JIGp8nsDK
+CwZvlu501j2y+yEZ4mUkNkbwe1U7bYd/KTexbsgrpzxfcqsPNGrs3te2k6m2Bkp1
+KS43rUKx9VvjIaYstx9akuCZmirTLWimCUOGTlsnqGLf5y806F+ifyHrBLkWa6ly
+cZGnGitFS/PzCR3gjoO1CNIDFU5bLhAMGFdC6lr7aHLVctTI2Q9O42PVCBf169nP
+10poZACmxQkuDvKVDRA6foNibTWbwHq2Ng//TORMdCts3BUf+cpXBX+zXj3qV3Nc
+CPgnQwLaeotV1iBhzrt/TtWmZI2eGGNB8oL2qA==
+-----END CERTIFICATE-----
