+++ sync/vault-commons/src/test/resources/client_cert.pem
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDTjCCAjagAwIBAgIEZYwSAzANBgkqhkiG9w0BAQsFADBpMQswCQYDVQQGEwJD
+TzELMAkGA1UECAwCQU4xETAPBgNVBAcMCE1FREVMTElOMRIwEAYDVQQKDAlsb2Nh
+bGhvc3QxEjAQBgNVBAsMCWxvY2FsaG9zdDESMBAGA1UEAwwJbG9jYWxob3N0MB4X
+DTIzMTIyNzEyMDEwN1oXDTI0MTIyNjEyMDEwN1owaTELMAkGA1UEBhMCQ08xCzAJ
+BgNVBAgMAkFOMREwDwYDVQQHDAhNRURFTExJTjESMBAGA1UECgwJbG9jYWxob3N0
+MRIwEAYDVQQLDAlsb2NhbGhvc3QxEjAQBgNVBAMMCWxvY2FsaG9zdDCCASIwDQYJ
+KoZIhvcNAQEBBQADggEPADCCAQoCggEBAMkh7JoVfFX20p0pqtRS8rRUl+70mpeu
+32fGoN5pA1AdqTRZfOoGKoHnVhAG6C/5UoD5GOStKLVdsXVEtaWkBnyVAD3Imdj7
+450dU9+LATgSQYJbTL11OcZBizeiu6DfZmBtYWvmpPDIAPsZr6R+kcUIrzH3ZQjU
+/BAfwCEmdQYwW7gWL456mTvefnMXcraoS4sOlBrH5FGXkb0ESIHuXMUsOrvg3cgN
+8S9R7huZQTDCio605KZW3hjfngBSAFTKbxxNI8nt05i5Gd+PMjpTndBUT4xYPp8Y
+igjL5IE0j9xY6IfvHqn/GD5GzYYIjjer2eCGmXfHDNtHAnA/wgdArEsCAwEAATAN
+BgkqhkiG9w0BAQsFAAOCAQEAKUcvHUTOL+rNr4DvdS3N3PSvVsnldgcYoBGZH0nf
+QtSWdbtLXfnhzgQPP+ndvL8ZWrYyZAv6+/fV7eQ569436nu+ebPz43JYiUsPReSg
+qzcJn95JYJn1nW6CKBhkkFGmcDMidm6u0eSNKZdKc+vPdlQFcmqEXjEv7gMHmPrE
+rCwXKvxQZ1DEFett56DFqJvnOTD0lBAGl6pyROqsF5JjJVUXgGcHGbXll3/CzgvK
+803fOq4kgu/BNPxy/KyC6Ztz5Qgr+7/zFC52nyRJilWPOJwU64fHu+yHaR9sI6bn
+NE3wY5gnwzKHWL2jotlFfSuoXNpBW11JSdzfDDH5GWhMDQ==
+-----END CERTIFICATE-----
