+++ sync/vault-sync/vault-sync.gradle
@@ -1,0 +1,13 @@
+dependencies {
+    api project(":secrets-manager-api")
+    api project(":vault-commons")
+    implementation "com.google.code.gson:gson:${gsonVersion}"
+    implementation "com.github.ben-manes.caffeine:caffeine:${cafeineVersion}"
+    testImplementation("com.squareup.okhttp3:mockwebserver:4.9.3")
+
+}
+
+ext {
+    artifactId = 'vault-sync'
+    artifactDescription = 'Secrets Manager connector for Vault'
+}
